requirements.txt
litellm
tqdm
datasets
python_dotenv
git+https://github.com/threewisemonkeys-as/athils.git